Finish notifications and other bug fixes #5

+++ src/main/kotlin/com/innopolis/innoqueue/controller/dto/TaskDTO.kt
@@ -7,5 +7,5 @@ data class TaskDTO(
     val taskId: Long,
 
     @JsonProperty("expenses")
-    val expenses: Int?
+    val expenses: Double?
 )
